+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+  * Enforce receive-side max message size #27
+
 # 0.7.1
 
   * Fix server-side rwnd initialization from INIT #28